+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - migrated SMS default URL (<https://sms.iex.ec>)
+- migrated from SMS deprecated apps secrets endpoint
 
 ## [8.13.0] 2025-01-21